requirements.txt
# Chess Transformer - pip dependencies
# (PyTorch is installed via conda separately for CUDA support)

# PGN parsing & move conversion to UCI
python-chess

# HuggingFace dataset download
huggingface_hub

# Parquet query with predicate pushdown (httpfs)
duckdb
pyarrow
pandas

# Vocabulary building (HuggingFace)
tokenizers

# Config parsing
pyyaml

# Model export & inference
onnx
onnxruntime

# Lichess bot API client
httpx
